requirements.txt
Django==6.0.1
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
psycopg2-binary==2.9.11
python-dotenv==1.2.1
django-cors-headers==4.9.0
pandas==2.3.3
numpy==2.4.1
scikit-learn==1.8.0
celery==5.5.3
redis==7.1.0
drf-spectacular==0.27.1

# ML Packages
xgboost==1.7.6
joblib==1.3.2
matplotlib==3.7.2
seaborn==0.12.2